requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
cycler==0.10.0
defusedxml==0.6.0
distlib==0.3.1
dj-rest-auth==2.1.3
Django==3.1.6
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
filelock==3.0.12
idna==2.10
importlib-metadata==3.4.0
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
oauthlib==3.1.0
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
pyparsing==2.4.5
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.2
zipp==3.4.0
